这是可以调试的版本。删除数据需要优化，插入数据还可以进一步优化

+++ sortchain_speedup/sortchain.h
@@ -15,7 +15,7 @@ extern "C" {
 #endif
 
 // user configure here
-typedef int schdat_t;       /* sort chain data */
+typedef float schdat_t;       /* sort chain data */
 #define SCH_NODES_TOTAL     (3999)
 
 #define SCH_TRUE    (1)
